Fix Firebase connectivity, data models, crash bugs & screen keep-on
- Add @IgnoreExtraProperties to all 7 data models (SofaStatus, DeviceStatus, Controls, ElectricalInfo, EnergyData, EnergyHistory, User, NotificationItem, HistoryItem) to prevent DatabaseException when ESP32 sends extra fields
- Change ElectricalInfo & EnergyData Float -> Double (Firebase stores numbers as Double, prevents type conversion crashes)
- Clear fake hardcoded DeviceStatus defaults (IP, MAC, signal) so UI shows real Firebase data
- Fix DashboardScreen: occupied time now uses lastOccupiedAt (was using lastEmptyAt), handle 0L timestamps
- Fix mode toggle: uses equalsIgnoreCase so 'auto'/'Auto'/'AUTO' all work correctly
- Fix DashboardViewModel: loadUserName() fetches real name from Firebase RTDB profile
- Fix FirebaseAuthManager: set displayName via UserProfileChangeRequest on registration so dashboard shows real user name; make database ref lazy
- Fix EnergyViewModel: use no-arg ElectricalInfo() default, Double cost calculation
- Fix EnergyScreen: change EnergySummaryCard Float->Double, remove unused Fill import
- Fix MainActivity: remove unused enableEdgeToEdge import
- Fix SmartSofaApp: remove crash-swallowing UncaughtExceptionHandler
- FLAG_KEEP_SCREEN_ON remains active to prevent auto screen-off

+++ app/src/main/java/com/example/smartsofa/ui/energy/EnergyScreen.kt
@@ -18,7 +18,6 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Path
-import androidx.compose.ui.graphics.drawscope.Fill
 import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
@@ -99,7 +98,7 @@ fun SummaryMiniCard(label: String, value: String, bgColor: Color, fgColor: Color
 }
 
 @Composable
-fun EnergySummaryCard(electricalInfo: ElectricalInfo, cost: Float) {
+fun EnergySummaryCard(electricalInfo: ElectricalInfo, cost: Double) {
     Card(shape = RoundedCornerShape(16.dp), modifier = Modifier.fillMaxWidth()) {
         Column(modifier = Modifier.padding(16.dp)) {
             Text("Summary", style = MaterialTheme.typography.titleMedium, fontWeight = FontWeight.Bold)
